Move unset_sslkeylogfile to General so startup skips requests
