+++ bleachbit/Network.py
@@ -59,27 +59,10 @@ else:
 # local imports
 from bleachbit import bleachbit_exe_path, APP_VERSION, ARCH_BITS, General, IS_LINUX, IS_NETBSD, IS_WINDOWS
 from bleachbit.FileUtilities import delete, open_for_overwrite
+from bleachbit.General import unset_sslkeylogfile
 from bleachbit.Language import get_active_language_code, get_text as _
 
 logger = logging.getLogger(__name__)
-
-
-def unset_sslkeylogfile(use_logger):
-    """Unset environment variable SSLKEYLOGFILE
-
-    Workaround for an OpenSSL crash before checking for updates.
-    https://github.com/bleachbit/bleachbit/issues/1826
-
-    Returns True if unset
-    """
-    if not IS_WINDOWS:
-        return False
-    if not os.environ.get('SSLKEYLOGFILE'):
-        return False
-    del os.environ['SSLKEYLOGFILE']
-    if use_logger:
-        logger.debug('The environment variable SSLKEYLOGFILE is not supported')
-    return True
 
 
 def download_url_to_fn(url, fn, expected_sha512=None, on_error=None,
